Add B3DM for tileset JSON creator. Minor cleanups.
